Refactor courses to consume YouTube API
Previously, you would need to manually add the publish date, image url
and channel id on the courses json. Not anymore. You just need the
course id. And I decided to keep the old json files for courses and
channels so that new developers who want to run techschool locally can
use them instead of having to generate a youtube api key to consume the
api